qbench: fix llama.cpp backend and cut reference logit VRAM
Two harness fixes required for the Qwen3.6-27B GGUF quality matrix on a
single 24 GB card. Local benchmark patches on top of upstream a0ef743;
not intended for submission.

engines.py:
- llama_cpp split-mode constants live under
  llama_cpp.llama_cpp.llama_split_mode.*, not at the package root, so
  building split_modes raised AttributeError before any model loaded.
- Pin n_batch/n_ubatch to 512.
- Llama.reset() only zeroes the Python-side token counter; the KV and
  recurrent state survive, so row r+1 starts at position 0 while
  sequence 0 still holds row r's positions and llama_decode rejects the
  batch with "inconsistent sequence positions". Clear the memory
  explicitly, plus the hybrid checkpoints, since rows are independent.

measure.py:
- Keep reference logits in their stored fp16 instead of upcasting to
  fp32. compute_kl_div casts to fp32 inside the kernel anyway, so the
  explicit upcast only doubled the reference footprint (3.8 GB/row at
  248320 vocab), which decided whether the pass fit in 24 GB.

+++ eval/qbench/measure.py
@@ -58,7 +58,10 @@ class DiffStats:
         # kld vs reference
         if self.ref_store:
             ref = load_tensor(os.path.join(self.ref_store, f"row_{r:06d}.safetensors"), "logits")
-            ref = ref.to(logits.device).float()
+            # Kept in the stored fp16: compute_kl_div casts to fp32 inside the kernel, so an
+            # explicit upcast here only doubles the reference's VRAM footprint (3.8 GB/row at
+            # this vocab size), which is the difference between fitting and not on 24 GB
+            ref = ref.to(logits.device)
             vs = min(self.vocab_size, logits.shape[-1], ref.shape[-1])
             kl = compute_kl_div(logits.squeeze(0), ref.squeeze(0), vs)
             self.kl_toks.append(kl.flatten().float().cpu())
